requirements.txt
# Dashboard için minimal paketler
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.5
python-socketio==5.10.0
simple-websocket==1.0.0
gunicorn==21.2.0
eventlet==0.33.3

# PostgreSQL
psycopg2-binary==2.9.9
python-dotenv==1.0.0
